requirements.txt
Django==2.1.2

djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
drf-writable-nested==0.5.1
django-ipware==2.1.0
coreapi==2.3.3

factory-boy==2.11.1
Faker==0.9.1
mock==2.0.0
pytest==3.8.2
pytest-cov==2.6.0
pytest-django==3.4.3
pytest-env==0.6.2
pytest-lazy-fixture==0.4.2
pytest-mock==1.10.0
